requirements.txt
requests
pydantic-settings
llama-cpp-python
colorama
bs4
ollama
fastapi
python-dotenv
uvicorn
diffusers==0.30.0
transformers==4.44.0
accelerate 
safetensors 
pillow
nh3
markdown
imap_tools
chromadb
wikipedia